+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Multiple positions for Hint text to render
--
+- `initialLastElementUpdated` prop to Form components
 
 ### Changed